VARIATION THREE: added for loop to randomize item placement, deleted all the useless code (i think), combined items into one check overlap
a lot happened here but for the game to even run I had to do all this at once or else I'd be calling a function that wasn't there and break everything. for loop loops thru the array and randomizes placement & checks if any of them are caught. a lot to fix i cant reset the game for some reason also nothing is constrained. thank you every single p5js example in the world for letting me come this far.