VARIATION THREE: i learned what a while loop is, and used it to make sure my items dont spawn overlapping
this took me forever and a million different tries. basically i have an overlapping condition that i set as true at the beginning even tho thats kinda a lie, so the while loop places all the items, but then a for loop checks if each item is overlapping with the same logic as my other overlap checks kinda and if the overlap is found to be true it throws u back to the while loop. pretty loopy. also i then did the same thing but way easier to make sure the bombs dont overlap w the items. now i have to make sure nothing overlaps the frog and the counter and i am so scared