VARIATION THREE: after HOURS no items overlap
very similar logic as what i had, but all the checks exist in the same spawn item function and i made sure nothing was interfering. while loop is back thank god was scared i learned that for nothing. I cried at a family function trying to get this to work.