VARIATION THREE: PROPERLY implemented the flashlight feature
i did this at 4 am last night but it was 4 am so i forgot to save it to the repository whatever anyways i made this black screen, then used some code from a p5 example mathilde sent me to make this flashlight effect with repeating circles that reveal the items when the lightsourse ovelaps with them using this custom alpha thing. took a while to integrate properly but it worked with some conditionals!