copied index code over to variation one file, used index page to create a level menu
I decided to organize my variations in different files. I found the benefits of this to be that my pages were shorter and easier to read and theres no interference between levels. I also got to test our using query selectors from my html and have my html and css files be more involved with the js! had to tweak a lot of the basic canvas settings cuz i was having some issues with layout, but everythings good and centered now!